Consolidate some test utilities
